Moved functions in main to selection.c and message.c selection.c contains: fileSelection and chunkSelection message.c contains: completedMergeMsg, completedSplitMsg, and errorMsg.
