+++ Makefile
@@ -2,7 +2,7 @@ CC = gcc
 CFLAGS = -Wall -Wextra
 TARGET = fsm
 LDFLAGS = -lncurses
-OBJS = main.o splitfile.o mergefile.o fileSize.o
+OBJS = main.o splitfile.o mergefile.o fileSize.o selection.o message.o
 all: $(TARGET)
 
 $(TARGET) : $(OBJS)
